requirements.txt
streamlit
python-dotenv
pydantic
google-genai